% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{dprior}
\alias{dprior}
\title{dprior}
\usage{
dprior(x, V, nu, alpha.mu = NULL, alpha.V = NULL, variable = 1)
}
\arguments{
\item{x}{vector of values for which the density is returned}

\item{V}{V in MCMCglmm prior specification}

\item{nu}{nu in MCMCglmm prior specification}

\item{alpha.mu}{alpha.mu in MCMCglmm prior specification}

\item{alpha.V}{alpha.V in MCMCglmm prior specification}

\item{variable}{when multivariate priors specified, which variable they apply to}
}
\description{
Density function for MCMCglmm priors
}
\details{
When V and nu only are specified, gives an inverse wishart prior, and when a covariance matrix is specified for V, it gives marginal prior on variance of specified variable (p12 and p72 of course notes).
When alpha.mu and alpha.V are specified, it returns a parameter expanded prior (p129 of course notes).
}
