% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcMCMCglmm.R
\name{mcMCMCglmm}
\alias{mcMCMCglmm}
\title{mcMCMCglmm}
\usage{
mcMCMCglmm(..., chains = 4, n.cores = 1, prior = NULL, verbose = FALSE)
}
\arguments{
\item{...}{arguments passed to MCMCglmm}

\item{chains}{number of chains}

\item{n.cores}{number of cores}

\item{prior}{prior argument to MCMCglmm}

\item{verbose}{verbose argument to MCMCglmm}
}
\description{
Function to run multi-chain MCMCglmm models in parallel
}
\details{
Pass the same arguements to MCMCglmm. It a smilar MCMCglmm is returned, the only difference is the slots taht previously had single posteriors for each parameter, have mcmc.list objects with multiple posteriors for each parameter (e.g. Sol and VCV)
}
\author{
Joel Pick - joel.l.pick@gmail.com
}
