% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_priors.R
\name{pp_plot}
\alias{pp_plot}
\title{pp_plot}
\usage{
pp_plot(x, prior, posterior, main = "", xlab = "", ylim = NULL, xlim = NULL)
}
\arguments{
\item{x}{vector of values for which the density is returned}

\item{prior}{vector of the prior density at each of x}

\item{posterior}{posterior distribution with which to compare the prior}

\item{main}{plot title}

\item{xlab}{label for x axis}

\item{ylim}{limits for y axis}

\item{xlim}{limits for x axis}
}
\description{
Function to plot prior and posterior
}
\details{
Where x is x sequence for the x values, and also is used as the breaks in the histogram, prior is the prior density at x (i.e. applying the prior density function to x), and posterior is your posterior distribution
}
