% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcMCMCglmm.R
\name{summary.mcMCMCglmm}
\alias{summary.mcMCMCglmm}
\title{summary.mcMCMCglmm}
\usage{
\method{summary}{mcMCMCglmm}(object, random = FALSE, ...)
}
\arguments{
\item{object}{mcMCMCglmm object}

\item{random}{logical: should the random effects be summarised}

\item{...}{Additional arguments}
}
\description{
Function to provide summary for mcMCMCglmm model, providing give summary stats across multiple chains. Gives the same output as summary.MCMCglmm, with the additional of Rhat
}
\details{
Provides the same output as MCMCglmm summary function with the addition of Rhat
}
